Change to float for more consistency

+++ res/handlers/newtrail.php
@@ -29,8 +29,8 @@
 
     $name = strip_tags($_POST["name"]);
     $description = strip_tags($_POST["description"]);
-    $lat = (double)$_POST["lat"];
-    $lng = (double)$_POST["lng"];
+    $lat = (float)$_POST["lat"];
+    $lng = (float)$_POST["lng"];
     $upload = $_FILES["img"];
     $image_type = image_type_to_mime_type(exif_imagetype($upload["tmp_name"]));
     $allowed_ext = array("image/jpeg", "image/png"); // Array of allowed image extensions
